+++ packages/@expo/config-plugins/src/ios/utils/Xcodeproj.ts
@@ -63,10 +63,7 @@ export function sanitizedName(name: string) {
 }
 
 function sanitizedNameForProjects(name: string) {
-  return name
-    .replace(/[\W_]+/g, '')
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '');
+  return name.normalize('NFD').replace(/[\W_]+/g, '');
 }
 
 // TODO: it's silly and kind of fragile that we look at app config to determine